src/requirements.txt
Flask
Pillow
matplotlib
numpy
paho-mqtt